# tensor metadata