--- a/tinygrad/engine/realize.py
+++ b/tinygrad/engine/realize.py
@@ -161,7 +161,7 @@ class ExecItem:
         ptm = (colored(f"{et*1e3:9.2f}ms", "yellow") if et > 0.01 else f"{et*1e6:9.2f}us") if et is not None else ""
         print(f"{colored(f'*** {self.prg.dname[:7]:7s} {GlobalCounters.kernel_count:4d}', 'magenta' if jit else ('green' if self.prg.first_run else None))} {self.prg.display_name+' '*(38-ansilen(self.prg.display_name))} arg {len(self.bufs):3d} mem {GlobalCounters.mem_used/1e9:5.2f} GB " +  # noqa: E501
               (str() if et is None else f"tm {ptm}/{GlobalCounters.time_sum_s*1e3:9.2f}ms ({op_estimate/((et or 1e-20)*1e9):8.2f} GFLOPS, {mem_estimate/((et or 1e-20)*1e9):7.2f} GB/s)" +  # noqa: E501
-               f"{[repr(m) if DEBUG >= 3 else str(m) for m in self.metadata] if self.metadata else ''}"))
+               f" {[repr(m) if DEBUG >= 3 else str(m) for m in self.metadata] if self.metadata else ''}"))
       self.prg.first_run = False
     return et
 
@@ -183,7 +183,8 @@ def lower_schedule_item(si:ScheduleItem) -> ExecItem:
 
 def lower_schedule(schedule:List[ScheduleItem]) -> Generator[ExecItem, None, None]:
   while len(schedule):
-    try: yield lower_schedule_item(si := schedule.pop(0))
+    si = schedule.pop(0)
+    try: yield lower_schedule_item(si)
     except Exception as e:
       if DEBUG >= 2:
         print(f"error lowering {si.ast[0].op}")