--- a/tinygrad/helpers.py
+++ b/tinygrad/helpers.py
@@ -106,7 +106,7 @@ WINO, THREEFRY, CAPTURING = ContextVar("WINO", 0), ContextVar("THREEFRY", 0), Co
 GRAPH, GRAPHPATH, SAVE_SCHEDULE, RING = ContextVar("GRAPH", 0), getenv("GRAPHPATH", "/tmp/net"), ContextVar("SAVE_SCHEDULE", 0), ContextVar("RING", 1)
 MULTIOUTPUT, PROFILE = ContextVar("MULTIOUTPUT", 1), ContextVar("PROFILE", 0)
 
-@dataclass
+@dataclass(frozen=True)
 class Metadata:
   name: str
   backward: bool = False