… test/test_tensor.py