like #4225 and #4197 and #3302

`traceback` is really slow, just grabbing a frame is relatively fast.

DEBUG=2 output:
```
*** AMD      829 r_512_11_11_32_2_2                     arg   2 mem 12.94 GB tm     74.60us/   349.07ms (  106.30 GFLOPS,  531.49 GB/s) ['max_pool2d']
*** AMD      830 r_128_11_11_32_4_2_2                   arg   3 mem 12.94 GB tm    362.12us/   349.43ms (   87.59 GFLOPS,  131.39 GB/s) ['max_pool2d bw']
*** AMD      831 r_64_16_3_8_3_3_256_3_4_3_3            arg   4 mem 12.94 GB tm   3679.48us/   353.11ms ( 3325.44 GFLOPS,    6.32 GB/s) ['conv2d', 'relu']
*** AMD      832 r_2_8_7_7_8_16_512_4_4_3_3             arg   4 mem 12.95 GB tm   5374.20us/   358.48ms ( 2754.03 GFLOPS,    4.93 GB/s) ['conv2d', 'relu']
*** AMD      833 r_2_8_5_5_8_16_512_4_4_3_3n1           arg   5 mem 12.95 GB tm   3012.84us/   361.50ms ( 2508.58 GFLOPS,    7.44 GB/s) ['conv2d', 'relu', 'relu bw', '__sub__', 'square bw', '__add__']
*** AMD      834 r_8_32_25_2_16_3_128_3_4_4             arg   3 mem 12.95 GB tm   3841.96us/   365.34ms ( 1965.08 GFLOPS,   10.99 GB/s) ['conv2d bw']
*** AMD      835 r_8_32_8_16_4_7_7_4                    arg   3 mem 12.95 GB tm    105.88us/   365.44ms (  272.96 GFLOPS,  399.86 GB/s) ['relu bw', 'conv2d bw']
```

error from #5269
```
error lowering BufferOps.STORE
tensor operations:
[ kaiming_uniform - __main__:18::__init__,
  kaiming_uniform - __main__:19::__init__,
  cat - __main__:39::test_bug,
  __rmul__ - __main__:12::optimal_mse,
  abs - __main__:12::optimal_mse,
  sum - __main__:12::optimal_mse]
```

~~about 10ms or 20% slower:~~ now 5ms or 10% slower:
without:
```
***** model forward in  52.90 ms
***** model forward in  52.82 ms
***** model forward in  53.45 ms
***** model forward in  52.53 ms
***** model forward in  53.60 ms
***** model forward in  53.68 ms
***** model forward in  52.81 ms
***** model forward in  52.45 ms
***** model forward in  52.90 ms
***** model forward in  53.62 ms
```
with:
```
***** model forward in  57.12 ms
***** model forward in  57.77 ms
***** model forward in  56.08 ms
***** model forward in  57.04 ms
***** model forward in  57.88 ms
***** model forward in  56.86 ms
***** model forward in  57.93 ms
***** model forward in  57.37 ms
***** model forward in  56.70 ms
***** model forward in  56.81 ms
```